db: widen the parse-time line counter to uint64_t end-to-end
The line count is not bounded by the sequence count (unsigned int):
one long wrapped record spans many lines, so a file can pass 2^32
lines while within every other limit, and error messages would then
report a wrapped line number. Everything downstream of the counter --
find_abundance()'s parameter, Seq_stats::missingabundance_lineno --
was already uint64_t; this widens the three places still at 32 bits:
Entry::lineno, parse_fasta()'s running counter, and
parse_sequence_body()'s reference parameter.

Entry stays 40 bytes: its two Index members align it to 8, so the
32-bit lineno paid the same 8 bytes in padding.

+++ src/utils/seq_index.hpp
@@ -44,7 +44,13 @@ struct Index {
 // One fasta record as captured during parsing: source line number
 // plus header and sequence locations in the database vectors.
 struct Entry {
-  unsigned int lineno {1U};
+  // 64-bit: the line count is not bounded by the sequence count -- one
+  // long wrapped record spans many lines -- and everything downstream
+  // (find_abundance, Seq_stats::missingabundance_lineno) is already
+  // uint64_t. Widening costs nothing here: the two Index members align
+  // the struct to 8 bytes, so a 32-bit lineno paid the same 8 bytes in
+  // padding.
+  uint64_t lineno {1};
   struct Index header;
   struct Index sequence;
 };
